Refactor EventsView to be a whole lot more readable and maintainable.
