+++ src/lib/eventUtils.js
@@ -1,15 +1,46 @@
 import moment from 'moment'
 import map from 'lodash/map'
 
-const addNewEvent = (events, key, event) => {
-  const oldEvents = events[key] || []
-  return {
-    ...events,
-    [key]: [...oldEvents, event]
-  }
+/**
+ *  Return true if an event either
+ *    a) does not have a time of date set
+ *    b) spans multiple days
+ */
+const isAlldayOrMultiday = ({ start, end, isAllDay }) => {
+  const duration = moment.duration(moment(end).diff(moment(start))).asHours()
+  return isAllDay || duration > 24
 }
 
-const dayRangeKeys = (title, startDate, endDate, isAllDay) => {
+// Return true if the event started the day before
+const continuesBefore = (start, date) =>
+  moment(start).isBefore(date, 'day')
+
+// Return true if the event continues the next day
+const continuesAfter = (end, date, isAllDay) => {
+  const endDateAdjusted = isAllDay
+    ? moment(end).subtract(3, 'hours')
+    : moment(end)
+
+    return endDateAdjusted.isAfter(date, 'day')
+}
+
+// Group events into an array of days
+const eventsIntoDayGroups = (events) => {
+  const grouped = events
+    .sort((a, b) => moment(a.start).diff(moment(b.start)))
+    .reduce((acc, event) => {
+      let newAcc = acc
+      const dates = dayRangeKeys(event.start, event.end, event.isAllDay)
+      const length = dates.length
+      dates.forEach((key, dayNumber) => {
+        newAcc = addNewEvent(newAcc, key, { dayNumber, length, ...event })
+      })
+      return newAcc
+    }, {})
+  return map(grouped, (events, date) => ({ events, date }))
+}
+
+const dayRangeKeys = (startDate, endDate, isAllDay) => {
   let now = moment(startDate)
 
   const adjustedEndDate = isAllDay
@@ -27,50 +58,12 @@ const dayRangeKeys = (title, startDate, endDate, isAllDay) => {
   return dates
 }
 
-const eventsIntoDayGroups = (events) => {
-  const grouped = events
-    .sort((a, b) => moment(a.start).diff(moment(b.start)))
-    .reduce((acc, event) => {
-      const key = moment(event.start).format('YYYY-MM-DD')
-      const oldEvents = acc[key] || []
-      return {
-        ...acc,
-        [key]: [...oldEvents, event]
-      }
-    }, {})
-  return map(grouped, (events, date) => ({ events, date }))
+const addNewEvent = (events, key, event) => {
+  const oldEvents = events[key] || []
+  return {
+    ...events,
+    [key]: [...oldEvents, event]
+  }
 }
 
-const eventsIntoDayGroupsAdjusted = (events) => {
-  const grouped = events
-    .sort((a, b) => moment(a.start).diff(moment(b.start)))
-    .reduce((acc, event) => {
-      let newAcc = acc
-      const dates = dayRangeKeys(event.title, event.start, event.end, event.isAllDay)
-      const length = dates.length
-      dates.forEach((key, dayNumber) => {
-        newAcc = addNewEvent(newAcc, key, { dayNumber, length, ...event })
-      })
-      return newAcc
-    }, {})
-  return map(grouped, (events, date) => ({ events, date }))
-}
-
-const isAlldayOrMultiday = ({ title, start, end, isAllDay }) => {
-  const duration = moment.duration(moment(end).diff(moment(start))).asHours()
-  return isAllDay || duration > 24
-}
-
-const continuesBefore = (start, date) =>
-  moment(start).isBefore(date, 'day')
-
-const continuesAfter = (end, date, isAllDay) => {
-  const endDateAdjusted = isAllDay
-    ? moment(end).subtract(3, 'hours')
-    : moment(end)
-  return endDateAdjusted.isAfter(date, 'day')
-}
-
-
-
-export { eventsIntoDayGroups, eventsIntoDayGroupsAdjusted, isAlldayOrMultiday, continuesBefore, continuesAfter }
+export { isAlldayOrMultiday, continuesBefore, continuesAfter, eventsIntoDayGroups }
